Add Parser, TaskList, Ui.java
